refactor(pipeline): extract guardrails.go from engine.go [#285]
Move budget enforcement, phase gating, rework routing, and regression
detection functions into a dedicated guardrails.go file to reduce the
size of engine.go and improve code organization.

Extracted functions: checkBudget, checkPhaseBudget, gatePhase, gateRework,
gateVerifyFail, gatePatchResult, handlePatchExhausted, downgradeToFollowUps,
detectRegression, extractFailingCriteria, and associated types (reworkVerdict,
regressionResult).
